requirements.txt
streamlit
Pillow
numpy
scikit-image
opencv-python-headless